# api-backend-nodejs/swagger.yaml
openapi: 3.0.3
info:
  title: CutQ Management API
  description: |
    Comprehensive API for CutQ (beauty business) booking and management system.

    ## Features
    - CutQ management with auto-increment display IDs
    - SEO-friendly URLs for better search indexing
    - Service booking and management with verification codes
    - Review and rating system
    - Organized file upload capabilities
    - Special offers and promotions
    - Social authentication (Google, Facebook)
    - Multi-role authentication (Admin/Salon Owner/Customer)

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## SEO-Friendly URLs
    The API supports SEO-optimized URLs for better search engine indexing:
    - CutQ URLs: `/salons/{salon-name}/{display-id}` (e.g., `/salons/beautiful-cuts/123`)
    - Service URLs: `/services/{category}/{service-id}` (e.g., `/services/hair/456`)
    - Backward compatibility maintained for legacy URLs

    ## Display IDs
    The API supports both UUID and auto-increment display IDs for user-friendly URLs:
    - UUIDs: `f12ebd4a-94b0-4162-bd7e-37858ea61d93`
    - Display IDs: `1`, `2`, `3`

  version: 1.2.0
  contact:
    name: CutQ Management API
    email: support@cutq.store
servers:
  - url: http://localhost:3002/api/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error message"
...
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        salonId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "Excellent service!"
        createdAt:
          type: string
          format: date-time
...
paths:
...
  /reviews/salon/{salonId}:
    get:
      tags:
        - Reviews
      summary: Get reviews for a salon
      description: Retrieve reviews for a specific salon with pagination
      parameters:
        - name: salonId
          in: path
          required: true
          description: Salon ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Number of reviews per page
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Reviews retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      reviews:
                        type: array
                        items:
                          $ref: '#/components/schemas/Review'
                      pagination:
                        type: object
                        properties:
                          page:
                            type: integer
                            example: 1
                          limit:
                            type: integer
                            example: 10
                          total:
                            type: integer
                            example: 50
                          pages:
                            type: integer
                            example: 5
                      averageRating:
                        type: number
                        example: 4.5
                      totalReviews:
                        type: integer
                        example: 50
  /reviews:
    post:
      tags:
        - Reviews
      summary: Create a new review
      description: Create a review for a completed booking (requires authentication)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - salonId
                - bookingId
                - rating
                - comment
              properties:
                salonId:
                  type: string
                  format: uuid
                bookingId:
                  type: string
                  format: uuid
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  example: "Excellent service! Highly recommend."
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Review'
                  message:
                    type: string
                    example: "Review created successfully"
        '400':
          description: Validation error or booking not eligible for review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Offer Endpoints
...